@@ -3,24 +3,35 @@
   "version": "0.1.0",
   "private": true,
   "devDependencies": {
-    "react-scripts": "0.6.1"
+    "cross-env": "^5.2.0",
+    "react-scripts": "3.1.1"
   },
   "dependencies": {
-    "history": "^4.3.0",
-    "marked": "^0.3.6",
+    "connected-react-router": "^6.5.2",
+    "history": "^4.9.0",
+    "marked": "^0.7.0",
     "prop-types": "^15.7.2",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
-    "react-redux": "^4.4.5",
-    "react-router": "^2.8.1",
-    "redux": "^3.6.0",
-    "superagent": "^3.7.0",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
+    "react-redux": "^7.1.0",
+    "react-router": "^5.0.1",
+    "react-router-dom": "^5.0.1",
+    "redux": "^4.0.4",
+    "redux-devtools-extension": "^2.13.8",
+    "redux-logger": "^3.0.6",
+    "superagent": "^5.1.0",
     "superagent-promise": "^1.1.0"
   },
   "scripts": {
-    "start": "PORT=5000 react-scripts start",
+    "start": "cross-env PORT=4100 react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
+    "test": "cross-env PORT=4100 react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
-  }
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
