@@ -4,6 +4,7 @@ import React from 'react';
 import agent from '../../agent';
 import { connect } from 'react-redux';
 import marked from 'marked';
+import { ARTICLE_PAGE_LOADED, ARTICLE_PAGE_UNLOADED } from '../../constants/actionTypes';
 
 const mapStateToProps = state => ({
   ...state.article,
@@ -12,16 +13,16 @@ const mapStateToProps = state => ({
 
 const mapDispatchToProps = dispatch => ({
   onLoad: payload =>
-    dispatch({ type: 'ARTICLE_PAGE_LOADED', payload }),
+    dispatch({ type: ARTICLE_PAGE_LOADED, payload }),
   onUnload: () =>
-    dispatch({ type: 'ARTICLE_PAGE_UNLOADED' })
+    dispatch({ type: ARTICLE_PAGE_UNLOADED })
 });
 
-class Article extends React.Component {
-  componentWillMount() {
+class Article extends React.PureComponent {
+  componentDidMount() {
     this.props.onLoad(Promise.all([
-      agent.Articles.get(this.props.params.id),
-      agent.Comments.forArticle(this.props.params.id)
+      agent.Articles.get(this.props.match.params.id),
+      agent.Comments.forArticle(this.props.match.params.id)
     ]));
   }
 
@@ -34,7 +35,7 @@ class Article extends React.Component {
       return null;
     }
 
-    const markup = { __html: marked(this.props.article.body) };
+    const markup = { __html: marked(this.props.article.body, { sanitize: true }) };
     const canModify = this.props.currentUser &&
       this.props.currentUser.username === this.props.article.author.username;
     return (
@@ -82,10 +83,10 @@ class Article extends React.Component {
 
           <div className="row">
             <CommentContainer
-                comments={this.props.comments || []}
-                errors={this.props.commentErrors}
-                slug={this.props.params.id}
-                currentUser={this.props.currentUser} />
+              comments={this.props.comments || []}
+              errors={this.props.commentErrors}
+              slug={this.props.match.params.id}
+              currentUser={this.props.currentUser} />
           </div>
         </div>
       </div>
