@@ -4,6 +4,11 @@ import React from 'react';
 import Tags from './Tags';
 import agent from '../../agent';
 import { connect } from 'react-redux';
+import {
+  HOME_PAGE_LOADED,
+  HOME_PAGE_UNLOADED,
+  APPLY_TAG_FILTER
+} from '../../constants/actionTypes';
 
 const Promise = global.Promise;
 
@@ -14,22 +19,22 @@ const mapStateToProps = state => ({
 });
 
 const mapDispatchToProps = dispatch => ({
-  onClickTag: (tag, payload) =>
-    dispatch({ type: 'APPLY_TAG_FILTER', tag, payload }),
-  onLoad: (tab, payload) =>
-    dispatch({ type: 'HOME_PAGE_LOADED', tab, payload }),
+  onClickTag: (tag, pager, payload) =>
+    dispatch({ type: APPLY_TAG_FILTER, tag, pager, payload }),
+  onLoad: (tab, pager, payload) =>
+    dispatch({ type: HOME_PAGE_LOADED, tab, pager, payload }),
   onUnload: () =>
-    dispatch({  type: 'HOME_PAGE_UNLOADED' })
+    dispatch({  type: HOME_PAGE_UNLOADED })
 });
 
-class Home extends React.Component {
-  componentWillMount() {
+class Home extends React.PureComponent {
+  componentDidMount() {
     const tab = this.props.token ? 'feed' : 'all';
     const articlesPromise = this.props.token ?
-      agent.Articles.feed() :
-      agent.Articles.all();
+      agent.Articles.feed :
+      agent.Articles.all;
 
-    this.props.onLoad(tab, Promise.all([agent.Tags.getAll(), articlesPromise]));
+    this.props.onLoad(tab, articlesPromise, Promise.all([agent.Tags.getAll(), articlesPromise()]));
   }
 
   componentWillUnmount() {
@@ -50,6 +55,7 @@ class Home extends React.Component {
               <div className="sidebar">
 
                 <p>Popular Tags</p>
+
                 <Tags
                   tags={this.props.tags}
                   onClickTag={this.props.onClickTag} />
