@@ -1,23 +1,29 @@
 import ArticleList from './ArticleList';
 import React from 'react';
-import { Link } from 'react-router';
+import { Link } from 'react-router-dom';
 import agent from '../agent';
 import { connect } from 'react-redux';
+import {
+  FOLLOW_USER,
+  UNFOLLOW_USER,
+  PROFILE_PAGE_LOADED,
+  PROFILE_PAGE_UNLOADED
+} from '../constants/actionTypes';
 
-const EditProfileSettings = props => {
+const EditProfileSettings = React.memo(props => {
   if (props.isUser) {
     return (
       <Link
-        to="settings"
+        to="/settings"
         className="btn btn-sm btn-outline-secondary action-btn">
         <i className="ion-gear-a"></i> Edit Profile Settings
       </Link>
     );
   }
   return null;
-};
+});
 
-const FollowUserButton = props => {
+const FollowUserButton = React.memo(props => {
   if (props.isUser) {
     return null;
   }
@@ -47,7 +53,7 @@ const FollowUserButton = props => {
       {props.user.following ? 'Unfollow' : 'Follow'} {props.user.username}
     </button>
   );
-};
+});
 
 const mapStateToProps = state => ({
   ...state.articleList,
@@ -57,23 +63,22 @@ const mapStateToProps = state => ({
 
 const mapDispatchToProps = dispatch => ({
   onFollow: username => dispatch({
-    type: 'FOLLOW_USER',
+    type: FOLLOW_USER,
     payload: agent.Profile.follow(username)
   }),
-  onLoad: payload => dispatch({ type: 'PROFILE_PAGE_LOADED', payload }),
-  onSetPage: (page, payload) => dispatch({ type: 'SET_PAGE', page, payload }),
+  onLoad: payload => dispatch({ type: PROFILE_PAGE_LOADED, payload }),
   onUnfollow: username => dispatch({
-    type: 'UNFOLLOW_USER',
+    type: UNFOLLOW_USER,
     payload: agent.Profile.unfollow(username)
   }),
-  onUnload: () => dispatch({ type: 'PROFILE_PAGE_UNLOADED' })
+  onUnload: () => dispatch({ type: PROFILE_PAGE_UNLOADED })
 });
 
-class Profile extends React.Component {
-  componentWillMount() {
+class Profile extends React.PureComponent {
+  componentDidMount() {
     this.props.onLoad(Promise.all([
-      agent.Profile.get(this.props.params.username),
-      agent.Articles.byAuthor(this.props.params.username)
+      agent.Profile.get(this.props.match.params.username),
+      agent.Articles.byAuthor(this.props.match.params.username)
     ]));
   }
 
@@ -87,7 +92,7 @@ class Profile extends React.Component {
         <li className="nav-item">
           <Link
             className="nav-link active"
-            to={`@${this.props.profile.username}`}>
+            to={`/@${this.props.profile.username}`}>
             My Articles
           </Link>
         </li>
@@ -95,17 +100,12 @@ class Profile extends React.Component {
         <li className="nav-item">
           <Link
             className="nav-link"
-            to={`@${this.props.profile.username}/favorites`}>
+            to={`/@${this.props.profile.username}/favorites`}>
             Favorited Articles
           </Link>
         </li>
       </ul>
     );
-  }
-
-  onSetPage(page) {
-    const promise = agent.Articles.byAuthor(this.props.profile.username, page);
-    this.props.onSetPage(page, promise);
   }
 
   render() {
@@ -116,8 +116,6 @@ class Profile extends React.Component {
 
     const isUser = this.props.currentUser &&
       this.props.profile.username === this.props.currentUser.username;
-
-    const onSetPage = page => this.onSetPage(page)
 
     return (
       <div className="profile-page">
@@ -154,10 +152,10 @@ class Profile extends React.Component {
               </div>
 
               <ArticleList
+                pager={this.props.pager}
                 articles={this.props.articles}
                 articlesCount={this.props.articlesCount}
-                currentPage={this.props.currentPage}
-                onSetPage={onSetPage} />
+                state={this.props.currentPage} />
             </div>
 
           </div>
