@@ -1,11 +1,12 @@
-import { Link } from 'react-router';
+import { Link } from 'react-router-dom';
 import React from 'react';
 import agent from '../../agent';
 import { connect } from 'react-redux';
+import { DELETE_ARTICLE } from '../../constants/actionTypes';
 
 const mapDispatchToProps = dispatch => ({
   onClickDelete: payload =>
-    dispatch({ type: 'DELETE_ARTICLE', payload })
+    dispatch({ type: DELETE_ARTICLE, payload })
 });
 
 const ArticleActions = props => {
@@ -37,4 +38,4 @@ const ArticleActions = props => {
   );
 };
 
-export default connect(() => ({}), mapDispatchToProps)(ArticleActions);
+export default connect(() => ({}), mapDispatchToProps)(React.memo(ArticleActions));
