@@ -1,17 +1,17 @@
 import ArticleActions from './ArticleActions';
-import { Link } from 'react-router';
-import React from 'react';
+import { Link } from 'react-router-dom';
+import React, {memo} from 'react';
 
 const ArticleMeta = props => {
   const article = props.article;
   return (
     <div className="article-meta">
-      <Link to={`@${article.author.username}`}>
+      <Link to={`/@${article.author.username}`}>
         <img src={article.author.image} alt={article.author.username} />
       </Link>
 
       <div className="info">
-        <Link to={`@${article.author.username}`} className="author">
+        <Link to={`/@${article.author.username}`} className="author">
           {article.author.username}
         </Link>
         <span className="date">
@@ -24,4 +24,4 @@ const ArticleMeta = props => {
   );
 };
 
-export default ArticleMeta;
+export default memo(ArticleMeta);
