@@ -1,6 +1,6 @@
 import CommentInput from './CommentInput';
 import CommentList from './CommentList';
-import { Link } from 'react-router';
+import { Link } from 'react-router-dom';
 import React from 'react';
 
 const CommentContainer = props => {
@@ -22,9 +22,9 @@ const CommentContainer = props => {
     return (
       <div className="col-xs-12 col-md-8 offset-md-2">
         <p>
-          <Link to="login">Sign in</Link>
+          <Link to="/login">Sign in</Link>
           &nbsp;or&nbsp;
-          <Link to="register">sign up</Link>
+          <Link to="/register">sign up</Link>
           &nbsp;to add comments on this article.
         </p>
 
@@ -37,4 +37,4 @@ const CommentContainer = props => {
   }
 };
 
-export default CommentContainer;
+export default React.memo(CommentContainer);
