@@ -19,4 +19,4 @@ const CommentList = props => {
   );
 };
 
-export default CommentList;
+export default React.memo(CommentList);
