@@ -2,7 +2,7 @@ import ArticlePreview from './ArticlePreview';
 import ListPagination from './ListPagination';
 import React from 'react';
 
-const ArticleList = props => {
+const ArticleList = React.memo(props => {
   if (!props.articles) {
     return (
       <div className="article-preview">Loading...</div>
@@ -28,11 +28,11 @@ const ArticleList = props => {
       }
 
       <ListPagination
+        pager={props.pager}
         articlesCount={props.articlesCount}
-        currentPage={props.currentPage}
-        onSetPage={props.onSetPage} />
+        currentPage={props.currentPage} />
     </div>
   );
-};
+});
 
 export default ArticleList;
