@@ -1,23 +1,24 @@
 import React from 'react';
-import { Link } from 'react-router';
+import { Link } from 'react-router-dom';
 import agent from '../agent';
 import { connect } from 'react-redux';
+import { ARTICLE_FAVORITED, ARTICLE_UNFAVORITED } from '../constants/actionTypes';
 
 const FAVORITED_CLASS = 'btn btn-sm btn-primary';
 const NOT_FAVORITED_CLASS = 'btn btn-sm btn-outline-primary';
 
 const mapDispatchToProps = dispatch => ({
   favorite: slug => dispatch({
-    type: 'ARTICLE_FAVORITED',
+    type: ARTICLE_FAVORITED,
     payload: agent.Articles.favorite(slug)
   }),
   unfavorite: slug => dispatch({
-    type: 'ARTICLE_UNFAVORITED',
+    type: ARTICLE_UNFAVORITED,
     payload: agent.Articles.unfavorite(slug)
   })
 });
 
-const ArticlePreview = props => {
+const ArticlePreview = React.memo(props => {
   const article = props.article;
   const favoriteButtonClass = article.favorited ?
     FAVORITED_CLASS :
@@ -73,6 +74,6 @@ const ArticlePreview = props => {
       </Link>
     </div>
   );
-}
+})
 
 export default connect(() => ({}), mapDispatchToProps)(ArticlePreview);
