@@ -1,7 +1,7 @@
-import { Link } from 'react-router';
 import React from 'react';
+import { Link } from 'react-router-dom';
 
-const LoggedOutView = props => {
+const LoggedOutView = React.memo(props => {
   if (!props.currentUser) {
     return (
       <ul className="nav navbar-nav pull-xs-right">
@@ -13,13 +13,13 @@ const LoggedOutView = props => {
         </li>
 
         <li className="nav-item">
-          <Link to="login" className="nav-link">
+          <Link to="/login" className="nav-link">
             Sign in
           </Link>
         </li>
 
         <li className="nav-item">
-          <Link to="register" className="nav-link">
+          <Link to="/register" className="nav-link">
             Sign up
           </Link>
         </li>
@@ -28,36 +28,36 @@ const LoggedOutView = props => {
     );
   }
   return null;
-};
+});
 
-const LoggedInView = props => {
+const LoggedInView = React.memo(props => {
   if (props.currentUser) {
     return (
       <ul className="nav navbar-nav pull-xs-right">
 
         <li className="nav-item">
-          <Link to="" className="nav-link">
+          <Link to="/" className="nav-link">
             Home
           </Link>
         </li>
 
         <li className="nav-item">
-          <Link to="editor" className="nav-link">
+          <Link to="/editor" className="nav-link">
             <i className="ion-compose"></i>&nbsp;New Post
           </Link>
         </li>
 
         <li className="nav-item">
-          <Link to="settings" className="nav-link">
+          <Link to="/settings" className="nav-link">
             <i className="ion-gear-a"></i>&nbsp;Settings
           </Link>
         </li>
 
         <li className="nav-item">
           <Link
-            to={`@${props.currentUser.username}`}
+            to={`/@${props.currentUser.username}`}
             className="nav-link">
-            <img src={props.currentUser.image || ''} className="user-pic" alt={props.currentUser.username} />
+            <img src={props.currentUser.image} className="user-pic" alt={props.currentUser.username} />
             {props.currentUser.username}
           </Link>
         </li>
@@ -67,9 +67,9 @@ const LoggedInView = props => {
   }
 
   return null;
-};
+});
 
-class Header extends React.Component {
+class Header extends React.PureComponent {
   render() {
     return (
       <nav className="navbar navbar-light">
