@@ -16,4 +16,4 @@ const Banner = ({ appName, token }) => {
   );
 };
 
-export default Banner;
+export default React.memo(Banner);
