@@ -1,49 +1,54 @@
-import { Link } from 'react-router';
-import ListErrors from './ListErrors';
-import React from 'react';
-import agent from '../agent';
-import { connect } from 'react-redux';
+import { Link } from 'react-router-dom'
+import ListErrors from './ListErrors'
+import React from 'react'
+import agent from '../agent'
+import { connect } from 'react-redux'
+import {
+  UPDATE_FIELD_AUTH,
+  LOGIN,
+  LOGIN_PAGE_UNLOADED
+} from '../constants/actionTypes'
 
-const mapStateToProps = state => ({ ...state.auth });
+const mapStateToProps = state => state.auth
 
 const mapDispatchToProps = dispatch => ({
   onChangeEmail: value =>
-    dispatch({ type: 'UPDATE_FIELD_AUTH', key: 'email', value }),
+    dispatch({ type: UPDATE_FIELD_AUTH, key: 'email', value }),
   onChangePassword: value =>
-    dispatch({ type: 'UPDATE_FIELD_AUTH', key: 'password', value }),
+    dispatch({ type: UPDATE_FIELD_AUTH, key: 'password', value }),
   onSubmit: (email, password) =>
-    dispatch({ type: 'LOGIN', payload: agent.Auth.login(email, password) }),
+    dispatch({ type: LOGIN, payload: agent.Auth.login(email, password) }),
   onUnload: () =>
-    dispatch({ type: 'LOGIN_PAGE_UNLOADED' })
-});
+    dispatch({ type: LOGIN_PAGE_UNLOADED })
+})
 
-class Login extends React.Component {
-  constructor() {
-    super();
-    this.changeEmail = ev => this.props.onChangeEmail(ev.target.value);
-    this.changePassword = ev => this.props.onChangePassword(ev.target.value);
+class Login extends React.PureComponent {
+  constructor () {
+    super()
+    this.changeEmail = ev => this.props.onChangeEmail(ev.target.value)
+    this.changePassword = ev => this.props.onChangePassword(ev.target.value)
     this.submitForm = (email, password) => ev => {
-      ev.preventDefault();
-      this.props.onSubmit(email, password);
-    };
+      ev.preventDefault()
+      this.props.onSubmit(email, password)
+    }
   }
 
-  componentWillUnmount() {
-    this.props.onUnload();
+  componentWillUnmount () {
+    this.props.onUnload()
   }
 
-  render() {
-    const email = this.props.email;
-    const password = this.props.password;
+  render () {
+    const email = this.props.email
+    const password = this.props.password
     return (
-      <div className="auth-page">
-        <div className="container page">
-          <div className="row">
+      <div className='auth-page'>
+        <div className='container page'>
+          <div className='row'>
 
-            <div className="col-md-6 offset-md-3 col-xs-12">
-              <h1 className="text-xs-center">Sign In</h1>
-              <p className="text-xs-center">
-                <Link to="register">
+            <div className='col-md-6 offset-md-3 col-xs-12'>
+              <h1 className='text-xs-center'>Sign In</h1>
+              <p className='text-xs-center'>
+                <Link to='/register'>
                   Need an account?
                 </Link>
               </p>
@@ -53,27 +58,29 @@ class Login extends React.Component {
               <form onSubmit={this.submitForm(email, password)}>
                 <fieldset>
 
-                  <fieldset className="form-group">
+                  <fieldset className='form-group'>
                     <input
-                      className="form-control form-control-lg"
-                      type="email"
-                      placeholder="Email"
+                      className='form-control form-control-lg'
+                      autoComplete='username'
+                      type='email'
+                      placeholder='Email'
                       value={email || ''}
                       onChange={this.changeEmail} />
                   </fieldset>
 
-                  <fieldset className="form-group">
+                  <fieldset className='form-group'>
                     <input
-                      className="form-control form-control-lg"
-                      type="password"
-                      placeholder="Password"
+                      className='form-control form-control-lg'
+                      type='password'
+                      autoComplete='current-password'
+                      placeholder='Password'
                       value={password || ''}
                       onChange={this.changePassword} />
                   </fieldset>
 
                   <button
-                    className="btn btn-lg btn-primary pull-xs-right"
-                    type="submit"
+                    className='btn btn-lg btn-primary pull-xs-right'
+                    type='submit'
                     disabled={this.props.inProgress}>
                     Sign in
                   </button>
@@ -85,8 +92,8 @@ class Login extends React.Component {
           </div>
         </div>
       </div>
-    );
+    )
   }
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(Login);
+export default connect(mapStateToProps, mapDispatchToProps)(Login)
