@@ -1,33 +1,20 @@
-import { Provider } from 'react-redux';
 import ReactDOM from 'react-dom';
+import { Provider } from 'react-redux';
 import React from 'react';
-import { Router, Route, IndexRoute, hashHistory } from 'react-router';
+import { store, history} from './store';
+
+import { Route, Switch } from 'react-router-dom';
+import { ConnectedRouter } from 'connected-react-router';
 
 import App from './components/App';
-import Article from './components/Article';
-import Editor from './components/Editor';
-import Home from './components/Home';
-import Login from './components/Login';
-import Profile from './components/Profile';
-import ProfileFavorites from './components/ProfileFavorites';
-import Register from './components/Register';
-import Settings from './components/Settings';
-import store from './store';
 
 ReactDOM.render((
   <Provider store={store}>
-    <Router history={hashHistory}>
-      <Route path="/" component={App}>
-        <IndexRoute component={Home} />
-        <Route path="login" component={Login} />
-        <Route path="register" component={Register} />
-        <Route path="settings" component={Settings} />
-        <Route path="article/:id" component={Article} />
-        <Route path="@:username" component={Profile} />
-        <Route path="@:username/favorites" component={ProfileFavorites} />
-        <Route path="editor" component={Editor} />
-        <Route path="editor/:slug" component={Editor} />
-      </Route>
-    </Router>
+    <ConnectedRouter history={history}>
+      <Switch>
+        <Route path="/" component={App} />
+      </Switch>
+    </ConnectedRouter>
   </Provider>
+
 ), document.getElementById('root'));
