@@ -1,42 +1,21 @@
+import {
+  ARTICLE_FAVORITED,
+  ARTICLE_UNFAVORITED,
+  SET_PAGE,
+  APPLY_TAG_FILTER,
+  HOME_PAGE_LOADED,
+  HOME_PAGE_UNLOADED,
+  CHANGE_TAB,
+  PROFILE_PAGE_LOADED,
+  PROFILE_PAGE_UNLOADED,
+  PROFILE_FAVORITES_PAGE_LOADED,
+  PROFILE_FAVORITES_PAGE_UNLOADED
+} from '../constants/actionTypes';
+
 export default (state = {}, action) => {
   switch (action.type) {
-    case 'HOME_PAGE_LOADED':
-      return {
-        ...state,
-        articles: action.payload[1].articles,
-        articlesCount: action.payload[1].articlesCount,
-        tab: action.tab,
-        currentPage: 0
-      };
-    case 'HOME_PAGE_UNLOADED':
-      return {};
-    case 'CHANGE_TAB':
-      return {
-        ...state,
-        articles: action.payload.articles,
-        articlesCount: action.payload.articlesCount,
-        tab: action.tab,
-        tag: null,
-        currentPage: 0
-      };
-    case 'SET_PAGE':
-      return {
-        ...state,
-        articles: action.payload.articles,
-        articlesCount: action.payload.articlesCount,
-        currentPage: action.page
-      };
-    case 'APPLY_TAG_FILTER':
-      return {
-        ...state,
-        articles: action.payload.articles,
-        articlesCount: action.payload.articlesCount,
-        tab: null,
-        tag: action.tag,
-        currentPage: 0
-      };
-    case 'ARTICLE_FAVORITED':
-    case 'ARTICLE_UNFAVORITED':
+    case ARTICLE_FAVORITED:
+    case ARTICLE_UNFAVORITED:
       return {
         ...state,
         articles: state.articles.map(article => {
@@ -50,16 +29,56 @@ export default (state = {}, action) => {
           return article;
         })
       };
-    case 'PROFILE_PAGE_LOADED':
-    case 'PROFILE_FAVORITES_PAGE_LOADED':
+    case SET_PAGE:
       return {
         ...state,
+        articles: action.payload.articles,
+        articlesCount: action.payload.articlesCount,
+        currentPage: action.page
+      };
+    case APPLY_TAG_FILTER:
+      return {
+        ...state,
+        pager: action.pager,
+        articles: action.payload.articles,
+        articlesCount: action.payload.articlesCount,
+        tab: null,
+        tag: action.tag,
+        currentPage: 0
+      };
+    case HOME_PAGE_LOADED:
+      return {
+        ...state,
+        pager: action.pager,
+        tags: action.payload[0].tags,
+        articles: action.payload[1].articles,
+        articlesCount: action.payload[1].articlesCount,
+        currentPage: 0,
+        tab: action.tab
+      };
+    case HOME_PAGE_UNLOADED:
+      return {};
+    case CHANGE_TAB:
+      return {
+        ...state,
+        pager: action.pager,
+        articles: action.payload.articles,
+        articlesCount: action.payload.articlesCount,
+        tab: action.tab,
+        currentPage: 0,
+        tag: null
+      };
+    case PROFILE_PAGE_LOADED:
+    case PROFILE_FAVORITES_PAGE_LOADED:
+      return {
+        ...state,
+        pager: action.pager,
         articles: action.payload[1].articles,
         articlesCount: action.payload[1].articlesCount,
         currentPage: 0
       };
-    case 'PROFILE_PAGE_UNLOADED':
-    case 'PROFILE_FAVORITES_PAGE_UNLOADED':
+    case PROFILE_PAGE_UNLOADED:
+    case PROFILE_FAVORITES_PAGE_UNLOADED:
       return {};
     default:
       return state;
