@@ -1,7 +1,16 @@
+import {
+  EDITOR_PAGE_LOADED,
+  EDITOR_PAGE_UNLOADED,
+  ARTICLE_SUBMITTED,
+  ASYNC_START,
+  ADD_TAG,
+  REMOVE_TAG,
+  UPDATE_FIELD_EDITOR
+} from '../constants/actionTypes';
 
 export default (state = {}, action) => {
   switch (action.type) {
-    case 'EDITOR_PAGE_LOADED':
+    case EDITOR_PAGE_LOADED:
       return {
         ...state,
         articleSlug: action.payload ? action.payload.article.slug : '',
@@ -11,31 +20,31 @@ export default (state = {}, action) => {
         tagInput: '',
         tagList: action.payload ? action.payload.article.tagList : []
       };
-    case 'EDITOR_PAGE_UNLOADED':
+    case EDITOR_PAGE_UNLOADED:
       return {};
-    case 'ARTICLE_SUBMITTED':
+    case ARTICLE_SUBMITTED:
       return {
         ...state,
         inProgress: null,
         errors: action.error ? action.payload.errors : null
       };
-    case 'ASYNC_START':
-      if (action.subtype === 'ARTICLE_SUBMITTED') {
+    case ASYNC_START:
+      if (action.subtype === ARTICLE_SUBMITTED) {
         return { ...state, inProgress: true };
       }
       break;
-    case 'ADD_TAG':
+    case ADD_TAG:
       return {
         ...state,
         tagList: state.tagList.concat([state.tagInput]),
         tagInput: ''
       };
-    case 'REMOVE_TAG':
+    case REMOVE_TAG:
       return {
         ...state,
         tagList: state.tagList.filter(tag => tag !== action.tag)
       };
-    case 'UPDATE_FIELD_EDITOR':
+    case UPDATE_FIELD_EDITOR:
       return { ...state, [action.key]: action.value };
     default:
       return state;
