@@ -1,27 +1,25 @@
-import { applyMiddleware, createStore, combineReducers } from 'redux';
+import { applyMiddleware, createStore } from 'redux';
+import { createLogger } from 'redux-logger'
+import { composeWithDevTools } from 'redux-devtools-extension/developmentOnly';
 import { promiseMiddleware, localStorageMiddleware } from './middleware';
-import article from './reducers/article';
-import articleList from './reducers/articleList';
-import auth from './reducers/auth';
-import common from './reducers/common';
-import editor from './reducers/editor';
-import home from './reducers/home';
-import profile from './reducers/profile';
-import settings from './reducers/settings';
+import createRootReducer from './reducer';
 
-const reducer = combineReducers({
-  article,
-  articleList,
-  auth,
-  common,
-  editor,
-  home,
-  profile,
-  settings
-});
+import { routerMiddleware } from 'connected-react-router';
+import { createBrowserHistory} from 'history';
 
-const middleware = applyMiddleware(promiseMiddleware, localStorageMiddleware);
+export const history = createBrowserHistory();
 
-const store = createStore(reducer, middleware);
+// Build the middleware for intercepting and dispatching navigation actions
+const myRouterMiddleware = routerMiddleware(history);
 
-export default store;
+const getMiddleware = () => {
+  if (process.env.NODE_ENV === 'production') {
+    return applyMiddleware(myRouterMiddleware, promiseMiddleware, localStorageMiddleware);
+  } else {
+    // Enable additional logging in non-production environments.
+    return applyMiddleware(myRouterMiddleware, promiseMiddleware, localStorageMiddleware, createLogger())
+  }
+};
+
+export const store = createStore(
+  createRootReducer(history), composeWithDevTools(getMiddleware()));
